FM1D2D-96 WIP Save/Load (commit with failing tests needed for working together with Luca)
Also a maptool added for generating 1D2D links 

Former-commit-id: f94c2f5b1c5fa19e5ab644e2878ccbad3eeb1c94
Former-commit-id: f70bf633a73517920777d47e18e5f3faa6be6208

+++ src/DeltaShell.NGHS.IO/Grid/UGrid1D2DLinksApi.cs
@@ -83,7 +83,7 @@ namespace DeltaShell.NGHS.IO.Grid
             var err = CheckArrayFormatWithFileDeclaration(mesh1DPointIdx, mesh2DFaceIdx, linkType, linkIds, linkLongNames, numberOf1D2DLinks);
             if( err != GridApiDataSet.GridConstants.NOERR) return err;
 
-            IntPtr intPtrMesh1Dindexes = Marshal.AllocCoTaskMem(Marshal.SizeOf(typeof(double)) * numberOf1D2DLinks);
+            IntPtr intPtrMesh1Dindexes = Marshal.AllocCoTaskMem(Marshal.SizeOf(typeof(int)) * numberOf1D2DLinks);
             IntPtr intPtrMesh2Dindexes = Marshal.AllocCoTaskMem(Marshal.SizeOf(typeof(int)) * numberOf1D2DLinks);
             IntPtr intPtrContactType = Marshal.AllocCoTaskMem(Marshal.SizeOf(typeof(int)) * numberOf1D2DLinks);
